refactor: flatten user feature namespaces
Move application classes into shared Actions, DTOs, Exceptions, and
Services namespaces.
