+++ app/Application/Actions/CreateUserAction.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Application\Users\Actions;
+namespace App\Application\Actions;
 
-use App\Application\Users\DTOs\CreateUserDTO;
-use App\Application\Users\DTOs\UserResponseDTO;
+use App\Application\DTOs\CreateUserDTO;
+use App\Application\DTOs\UserResponseDTO;
 use InvalidArgumentException;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
