+++ app/Application/Actions/ShowUserAction.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Application\Users\Actions;
+namespace App\Application\Actions;
 
-use App\Application\Users\DTOs\UserResponseDTO;
-use App\Application\Users\Exceptions\UserNotFoundException;
+use App\Application\DTOs\UserResponseDTO;
+use App\Application\Exceptions\UserNotFoundException;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
