+++ app/Application/DTOs/UserResponseDTO.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Application\Users\DTOs;
+namespace App\Application\DTOs;
 
 use App\Domain\User\User;
 use JsonSerializable;
