+++ tests/Unit/Application/DTOs/UserResponseDTOTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Application\Users\DTOs;
+namespace Tests\Unit\Application\DTOs;
 
-use App\Application\Users\DTOs\UserResponseDTO;
+use App\Application\DTOs\UserResponseDTO;
 use App\Domain\User\User;
 use PHPUnit\Framework\TestCase;
